Add $YAML::QuoteNumericStrings  setting.
This allows strings that merely /resemble/ numbers ( as opposed to being
provided to perl as a literal number ) being encoded safely as
strings.

This is for feature compatibility with YAML::XS's feature by the same
name.

NOTE: It is presently *disabled* by default as enabling by default has
too many knock-on side-effects and break tests.

For instance, any test with a numeric key stored in a hash
gets those hash keys quoted, because storing it in a hash
converts it from a number to a string.